Set up VR functionality by updating Unity editor
I have a functioning VR rig along with TP pads setup in the scene. There should also be a few mesh colliders on certain objects to prevent teleporting through walls and such. There is also text that should pop up when an object is looked at while on a certain pad, but that is not fully functional at the moment.