Updated unity editor version to 2020.3.26f1. Reimplemented all VR functionality, nothing broke.
As far as I can tell, everything looks the same, nothing was conflicted, the only different between this and my branch now is the hierarchy is much cleaner in my branch. In terms of visuals and function though, everything seems good. Just remember to install the correct editor version and include the android stuff during installation. If there's trouble, I should be able to talk you through it, just ask on discord.